+++ src/main/java/duke/Storage.java
@@ -1,5 +1,16 @@
-import java.io.*;
+package duke;
+
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.Todo;
+
 import java.util.Scanner;
+
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileWriter;
+import java.io.IOException;
 
 public class Storage {
     protected static String pwd;
