+++ src/main/java/duke/exception/BlankCommandException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class BlankCommandException extends DukeException {
 
     public BlankCommandException() {
