+++ src/main/java/duke/exception/IncompleteCommandException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class IncompleteCommandException extends DukeException {
     public IncompleteCommandException(String command) {
         super("OOPS!!! The description of a " + command + " cannot be empty.");
